# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maodo Services API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/services:
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdateInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, price, delay, description, icon, active]
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: integer
        delay:
          type: string
        description:
          type: string
        icon:
          type: string
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
    ServiceInput:
      type: object
      required: [id, name, description, price, delay]
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 5
        price:
          type: integer
          minimum: 0
        delay:
          type: string
        icon:
          type: string
    ServiceUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 5
        price:
          type: integer
          minimum: 0
        delay:
          type: string
        icon:
          type: string
        active:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...